refactor: clean up error types
