+++ halo2_frontend/src/dev/gates.rs
@@ -31,8 +31,6 @@ struct Gate {
 /// use halo2_frontend::{
 ///     circuit::{Layouter, SimpleFloorPlanner},
 ///     dev::CircuitGates,
-/// };
-/// use halo2_common::{
 ///     plonk::{Circuit, ConstraintSystem, Error},
 /// };
 /// use halo2curves::pasta::pallas;
