+++ halo2_proofs/examples/simple-example.rs
@@ -3,7 +3,7 @@ use std::marker::PhantomData;
 use halo2_proofs::{
     arithmetic::Field,
     circuit::{AssignedCell, Chip, Layouter, Region, SimpleFloorPlanner, Value},
-    plonk::{Advice, Circuit, Column, ConstraintSystem, Error, Fixed, Instance, Selector},
+    plonk::{Advice, Circuit, Column, ConstraintSystem, ErrorFront, Fixed, Instance, Selector},
     poly::Rotation,
 };
 
@@ -13,10 +13,18 @@ trait NumericInstructions<F: Field>: Chip<F> {
     type Num;
 
     /// Loads a number into the circuit as a private input.
-    fn load_private(&self, layouter: impl Layouter<F>, a: Value<F>) -> Result<Self::Num, Error>;
+    fn load_private(
+        &self,
+        layouter: impl Layouter<F>,
+        a: Value<F>,
+    ) -> Result<Self::Num, ErrorFront>;
 
     /// Loads a number into the circuit as a fixed constant.
-    fn load_constant(&self, layouter: impl Layouter<F>, constant: F) -> Result<Self::Num, Error>;
+    fn load_constant(
+        &self,
+        layouter: impl Layouter<F>,
+        constant: F,
+    ) -> Result<Self::Num, ErrorFront>;
 
     /// Returns `c = a * b`.
     fn mul(
@@ -24,7 +32,7 @@ trait NumericInstructions<F: Field>: Chip<F> {
         layouter: impl Layouter<F>,
         a: Self::Num,
         b: Self::Num,
-    ) -> Result<Self::Num, Error>;
+    ) -> Result<Self::Num, ErrorFront>;
 
     /// Exposes a number as a public input to the circuit.
     fn expose_public(
@@ -32,7 +40,7 @@ trait NumericInstructions<F: Field>: Chip<F> {
         layouter: impl Layouter<F>,
         num: Self::Num,
         row: usize,
-    ) -> Result<(), Error>;
+    ) -> Result<(), ErrorFront>;
 }
 // ANCHOR_END: instructions
 
@@ -152,7 +160,7 @@ impl<F: Field> NumericInstructions<F> for FieldChip<F> {
         &self,
         mut layouter: impl Layouter<F>,
         value: Value<F>,
-    ) -> Result<Self::Num, Error> {
+    ) -> Result<Self::Num, ErrorFront> {
         let config = self.config();
 
         layouter.assign_region(
@@ -169,7 +177,7 @@ impl<F: Field> NumericInstructions<F> for FieldChip<F> {
         &self,
         mut layouter: impl Layouter<F>,
         constant: F,
-    ) -> Result<Self::Num, Error> {
+    ) -> Result<Self::Num, ErrorFront> {
         let config = self.config();
 
         layouter.assign_region(
@@ -187,7 +195,7 @@ impl<F: Field> NumericInstructions<F> for FieldChip<F> {
         mut layouter: impl Layouter<F>,
         a: Self::Num,
         b: Self::Num,
-    ) -> Result<Self::Num, Error> {
+    ) -> Result<Self::Num, ErrorFront> {
         let config = self.config();
 
         layouter.assign_region(
@@ -223,7 +231,7 @@ impl<F: Field> NumericInstructions<F> for FieldChip<F> {
         mut layouter: impl Layouter<F>,
         num: Self::Num,
         row: usize,
-    ) -> Result<(), Error> {
+    ) -> Result<(), ErrorFront> {
         let config = self.config();
 
         layouter.constrain_instance(num.0.cell(), config.instance, row)
@@ -272,7 +280,7 @@ impl<F: Field> Circuit<F> for MyCircuit<F> {
         &self,
         config: Self::Config,
         mut layouter: impl Layouter<F>,
-    ) -> Result<(), Error> {
+    ) -> Result<(), ErrorFront> {
         let field_chip = FieldChip::<F>::construct(config);
 
         // Load our private values into the circuit.
