+++ halo2_proofs/examples/vector-ops-unblinded.rs
@@ -33,7 +33,7 @@ trait NumericInstructions<F: Field>: Chip<F> {
         &self,
         layouter: impl Layouter<F>,
         a: &[Value<F>],
-    ) -> Result<Vec<Self::Num>, Error>;
+    ) -> Result<Vec<Self::Num>, ErrorFront>;
 
     /// Returns `c = a * b`. The caller is responsible for ensuring that `a.len() == b.len()`.
     fn mul(
@@ -41,7 +41,7 @@ trait NumericInstructions<F: Field>: Chip<F> {
         layouter: impl Layouter<F>,
         a: &[Self::Num],
         b: &[Self::Num],
-    ) -> Result<Vec<Self::Num>, Error>;
+    ) -> Result<Vec<Self::Num>, ErrorFront>;
 
     /// Returns `c = a + b`. The caller is responsible for ensuring that `a.len() == b.len()`.
     fn add(
@@ -49,7 +49,7 @@ trait NumericInstructions<F: Field>: Chip<F> {
         layouter: impl Layouter<F>,
         a: &[Self::Num],
         b: &[Self::Num],
-    ) -> Result<Vec<Self::Num>, Error>;
+    ) -> Result<Vec<Self::Num>, ErrorFront>;
 
     /// Exposes a number as a public input to the circuit.
     fn expose_public(
@@ -57,7 +57,7 @@ trait NumericInstructions<F: Field>: Chip<F> {
         layouter: impl Layouter<F>,
         num: &Self::Num,
         row: usize,
-    ) -> Result<(), Error>;
+    ) -> Result<(), ErrorFront>;
 }
 // ANCHOR_END: instructions
 
@@ -204,7 +204,7 @@ impl<F: Field> NumericInstructions<F> for MultChip<F> {
         &self,
         mut layouter: impl Layouter<F>,
         values: &[Value<F>],
-    ) -> Result<Vec<Self::Num>, Error> {
+    ) -> Result<Vec<Self::Num>, ErrorFront> {
         let config = self.config();
 
         layouter.assign_region(
@@ -228,7 +228,7 @@ impl<F: Field> NumericInstructions<F> for MultChip<F> {
         _: impl Layouter<F>,
         _: &[Self::Num],
         _: &[Self::Num],
-    ) -> Result<Vec<Self::Num>, Error> {
+    ) -> Result<Vec<Self::Num>, ErrorFront> {
         panic!("Not implemented")
     }
 
@@ -237,7 +237,7 @@ impl<F: Field> NumericInstructions<F> for MultChip<F> {
         mut layouter: impl Layouter<F>,
         a: &[Self::Num],
         b: &[Self::Num],
-    ) -> Result<Vec<Self::Num>, Error> {
+    ) -> Result<Vec<Self::Num>, ErrorFront> {
         let config = self.config();
         assert_eq!(a.len(), b.len());
 
@@ -271,7 +271,7 @@ impl<F: Field> NumericInstructions<F> for MultChip<F> {
         mut layouter: impl Layouter<F>,
         num: &Self::Num,
         row: usize,
-    ) -> Result<(), Error> {
+    ) -> Result<(), ErrorFront> {
         let config = self.config();
 
         layouter.constrain_instance(num.0.cell(), config.instance, row)
@@ -286,7 +286,7 @@ impl<F: Field> NumericInstructions<F> for AddChip<F> {
         &self,
         mut layouter: impl Layouter<F>,
         values: &[Value<F>],
-    ) -> Result<Vec<Self::Num>, Error> {
+    ) -> Result<Vec<Self::Num>, ErrorFront> {
         let config = self.config();
 
         layouter.assign_region(
@@ -310,7 +310,7 @@ impl<F: Field> NumericInstructions<F> for AddChip<F> {
         _: impl Layouter<F>,
         _: &[Self::Num],
         _: &[Self::Num],
-    ) -> Result<Vec<Self::Num>, Error> {
+    ) -> Result<Vec<Self::Num>, ErrorFront> {
         panic!("Not implemented")
     }
 
@@ -319,7 +319,7 @@ impl<F: Field> NumericInstructions<F> for AddChip<F> {
         mut layouter: impl Layouter<F>,
         a: &[Self::Num],
         b: &[Self::Num],
-    ) -> Result<Vec<Self::Num>, Error> {
+    ) -> Result<Vec<Self::Num>, ErrorFront> {
         let config = self.config();
         assert_eq!(a.len(), b.len());
 
@@ -353,7 +353,7 @@ impl<F: Field> NumericInstructions<F> for AddChip<F> {
         mut layouter: impl Layouter<F>,
         num: &Self::Num,
         row: usize,
-    ) -> Result<(), Error> {
+    ) -> Result<(), ErrorFront> {
         let config = self.config();
 
         layouter.constrain_instance(num.0.cell(), config.instance, row)
@@ -395,7 +395,7 @@ impl<F: Field> Circuit<F> for MulCircuit<F> {
         &self,
         config: Self::Config,
         mut layouter: impl Layouter<F>,
-    ) -> Result<(), Error> {
+    ) -> Result<(), ErrorFront> {
         let field_chip = MultChip::<F>::construct(config);
 
         // Load our unblinded values into the circuit.
@@ -448,7 +448,7 @@ impl<F: Field> Circuit<F> for AddCircuit<F> {
         &self,
         config: Self::Config,
         mut layouter: impl Layouter<F>,
-    ) -> Result<(), Error> {
+    ) -> Result<(), ErrorFront> {
         let field_chip = AddChip::<F>::construct(config);
 
         // Load our unblinded values into the circuit.
