+++ halo2_proofs/src/lib.rs
@@ -15,8 +15,8 @@ pub mod plonk;
 
 /// Traits and structs for implementing circuit components.
 pub mod circuit {
-    pub use halo2_common::circuit::floor_planner;
-    pub use halo2_common::circuit::{
+    pub use halo2_frontend::circuit::floor_planner;
+    pub use halo2_frontend::circuit::{
         AssignedCell, Cell, Chip, Layouter, Region, SimpleFloorPlanner, Value,
     };
 }
