+++ halo2_proofs/tests/plonk_api.rs
@@ -8,8 +8,8 @@ use halo2_proofs::circuit::{Cell, Layouter, SimpleFloorPlanner, Value};
 use halo2_proofs::dev::MockProver;
 use halo2_proofs::plonk::{
     create_proof as create_plonk_proof, keygen_pk, keygen_vk, verify_proof as verify_plonk_proof,
-    Advice, Assigned, Circuit, Column, ConstraintSystem, Error, Fixed, ProvingKey, TableColumn,
-    VerifyingKey,
+    Advice, Assigned, Circuit, Column, ConstraintSystem, Error, ErrorBack, ErrorFront, Fixed,
+    ProvingKey, TableColumn, VerifyingKey,
 };
 use halo2_proofs::poly::commitment::{CommitmentScheme, ParamsProver, Prover, Verifier};
 use halo2_proofs::poly::Rotation;
@@ -51,25 +51,34 @@ fn plonk_api() {
             &self,
             layouter: &mut impl Layouter<FF>,
             f: F,
-        ) -> Result<(Cell, Cell, Cell), Error>
+        ) -> Result<(Cell, Cell, Cell), ErrorFront>
         where
             F: FnMut() -> Value<(Assigned<FF>, Assigned<FF>, Assigned<FF>)>;
         fn raw_add<F>(
             &self,
             layouter: &mut impl Layouter<FF>,
             f: F,
-        ) -> Result<(Cell, Cell, Cell), Error>
+        ) -> Result<(Cell, Cell, Cell), ErrorFront>
         where
             F: FnMut() -> Value<(Assigned<FF>, Assigned<FF>, Assigned<FF>)>;
-        fn copy(&self, layouter: &mut impl Layouter<FF>, a: Cell, b: Cell) -> Result<(), Error>;
-        fn public_input<F>(&self, layouter: &mut impl Layouter<FF>, f: F) -> Result<Cell, Error>
+        fn copy(
+            &self,
+            layouter: &mut impl Layouter<FF>,
+            a: Cell,
+            b: Cell,
+        ) -> Result<(), ErrorFront>;
+        fn public_input<F>(
+            &self,
+            layouter: &mut impl Layouter<FF>,
+            f: F,
+        ) -> Result<Cell, ErrorFront>
         where
             F: FnMut() -> Value<FF>;
         fn lookup_table(
             &self,
             layouter: &mut impl Layouter<FF>,
             values: &[FF],
-        ) -> Result<(), Error>;
+        ) -> Result<(), ErrorFront>;
     }
 
     #[derive(Clone)]
@@ -97,7 +106,7 @@ fn plonk_api() {
             &self,
             layouter: &mut impl Layouter<FF>,
             mut f: F,
-        ) -> Result<(Cell, Cell, Cell), Error>
+        ) -> Result<(Cell, Cell, Cell), ErrorFront>
         where
             F: FnMut() -> Value<(Assigned<FF>, Assigned<FF>, Assigned<FF>)>,
         {
@@ -151,7 +160,7 @@ fn plonk_api() {
             &self,
             layouter: &mut impl Layouter<FF>,
             mut f: F,
-        ) -> Result<(Cell, Cell, Cell), Error>
+        ) -> Result<(Cell, Cell, Cell), ErrorFront>
         where
             F: FnMut() -> Value<(Assigned<FF>, Assigned<FF>, Assigned<FF>)>,
         {
@@ -211,7 +220,7 @@ fn plonk_api() {
             layouter: &mut impl Layouter<FF>,
             left: Cell,
             right: Cell,
-        ) -> Result<(), Error> {
+        ) -> Result<(), ErrorFront> {
             layouter.assign_region(
                 || "copy",
                 |mut region| {
@@ -220,7 +229,11 @@ fn plonk_api() {
                 },
             )
         }
-        fn public_input<F>(&self, layouter: &mut impl Layouter<FF>, mut f: F) -> Result<Cell, Error>
+        fn public_input<F>(
+            &self,
+            layouter: &mut impl Layouter<FF>,
+            mut f: F,
+        ) -> Result<Cell, ErrorFront>
         where
             F: FnMut() -> Value<FF>,
         {
@@ -243,7 +256,7 @@ fn plonk_api() {
             &self,
             layouter: &mut impl Layouter<FF>,
             values: &[FF],
-        ) -> Result<(), Error> {
+        ) -> Result<(), ErrorFront> {
             layouter.assign_table(
                 || "",
                 |mut table| {
@@ -369,7 +382,7 @@ fn plonk_api() {
             &self,
             config: PlonkConfig,
             mut layouter: impl Layouter<F>,
-        ) -> Result<(), Error> {
+        ) -> Result<(), ErrorFront> {
             let cs = StandardPlonk::new(config);
 
             let _ = cs.public_input(&mut layouter, || Value::known(F::ONE + F::ONE))?;
@@ -421,9 +434,9 @@ fn plonk_api() {
             let much_too_small_params= <$scheme as CommitmentScheme>::ParamsProver::new(1);
             assert_matches!(
                 keygen_vk(&much_too_small_params, &empty_circuit),
-                Err(Error::NotEnoughRowsAvailable {
+                Err(Error::Backend(ErrorBack::NotEnoughRowsAvailable {
                     current_k,
-                }) if current_k == 1
+                })) if current_k == 1
             );
 
             // Check that we get an error if we try to initialize the proving key with a value of
@@ -431,9 +444,9 @@ fn plonk_api() {
             let slightly_too_small_params = <$scheme as CommitmentScheme>::ParamsProver::new(K-1);
             assert_matches!(
                 keygen_vk(&slightly_too_small_params, &empty_circuit),
-                Err(Error::NotEnoughRowsAvailable {
+                Err(Error::Backend(ErrorBack::NotEnoughRowsAvailable {
                     current_k,
-                }) if current_k == K - 1
+                })) if current_k == K - 1
             );
         }};
     }
